Move load_json_dump to manage.py command
