--- scraper/templates/scraper/json_load.html
@@ -1,8 +1,0 @@
-<h1>JSON load page</h1>
-<h2> Log: </h2>
-<p>
-{% for line in log %}
-	{{ line }}
-	<br/>
-{% endfor %}
-</p>
